Zmiana z liczb calkowitych na rzeczywiste

+++ 02-petle/zad6.c
@@ -1,20 +1,20 @@
 #include <stdio.h>
 
-int liczby[6];
-int najwieksza;
+float liczby[6];
+float najwieksza;
 int i=1, j;
 
 int main() {
   printf("Podj szesc liczb calkowitych:\n");
   while (i<=6) {
     printf("liczba [%i]=", i);
-    scanf("%i", &liczby[i-1]);
+    scanf("%f", &liczby[i-1]);
     i++;
   }
 
   printf("\nWprowadzone liczby to:\n");
   for (j = 0; j < 6; j++) {
-    printf("%i ", liczby[j]);
+    printf("%f ", liczby[j]);
   }
   najwieksza = liczby[0];
   for (j = 0; j < 6; j++) {
@@ -22,5 +22,5 @@ int main() {
       najwieksza=liczby[j];
     }
   }
-  printf("\n\nNajewieksza liczba ze zbioru to: %i\n", najwieksza);
+  printf("\n\nNajewieksza liczba ze zbioru to: %f\n", najwieksza);
 }
